pyproject.toml
[tool.poetry]
name = "googleforms"
version = "0.1.0"
description = ""
authors = ["cheerio-pixel <frailin300@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9, <3.13"
google-auth-oauthlib = "^1.1.0"
google-auth = "^2.23.3"
requests = "^2.31.0"
google-api-python-client = "^2.104.0"
oauth2client = "^4.1.3"
sidetable = "^0.9.1"
pandas = "^2.1.2"
jinja2 = "^3.1.2"
matplotlib = "^3.8.1"
ipython = "^8.17.2"
dataframe-image = "^0.2.2"
matplotx = "^0.3.10"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.3"
schedule = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
